move hif functionality to convert
